resolve missing data in int

+++ pytidycensus/utils.py
@@ -830,7 +830,7 @@ def process_census_data(
     # Convert numeric columns
     for var in variables:
         if var in df.columns:
-            df[var] = pd.to_numeric(df[var], errors="coerce")
+            df[var] = pd.to_numeric(df[var], errors="coerce").astype("Int64")
 
     # Create GEOID from geography columns
     geo_cols = [
